# Conflicts:
#	browserassets/tgui/tgui.bundle.js
#	code/WorkInProgress/artemis/artemis.dm
#	code/WorkInProgress/artemis/artemis_movement_controller.dm
#	code/WorkInProgress/artemis/background_ship.dm
#	code/WorkInProgress/artemis/background_stars.dm
#	code/WorkInProgress/artemis/debug.dm
#	code/WorkInProgress/artemis/flight_computer.dm
#	code/WorkInProgress/artemis/flight_computer_hud.dm
#	code/WorkInProgress/artemis/galactic_object.dm
#	code/WorkInProgress/artemis/galactic_object/galactic_objects.dm
#	code/WorkInProgress/artemis/galactic_object/navigation_satellite.dm
#	code/WorkInProgress/artemis/planetary_teleporter.dm
#	code/datums/gamemodes/artemis_freeroam.dm
#	code/modules/telescience/telescope/telescopeMisc.dm
#	code/modules/worldgen/GenerateMining.dm
#	maps/config/map.dm

+++ code/mob/living/silicon/hivebot.dm
@@ -945,6 +945,8 @@ Frequency:
 		if(!bioHolder)
 			bioHolder = new/datum/bioHolder( src )
 		SPAWN(0.5 SECONDS)
+			if (QDELETED(src)) //bleh
+				return
 			if (src.module)
 				qdel(src.module)
 			if (ticker?.mode && istype(ticker.mode, /datum/game_mode/construction))
